numcodecs_wasm_host

Struct WasmCodecComponent

pub struct WasmCodecComponent { /* private fields */ }

WebAssembly component that exports the numcodecs:abc/codec interface.

WasmCodecComponent does not implement the DynCodecType trait itself so that it can expose un-opinionated bindings. However, it provides methods with that can be used to implement the trait on a wrapper.

Implementations

impl WasmCodecComponent

pub fn new( ctx: impl AsContextMut, instance: Instance, ) -> Result<Self, RuntimeError>

Import the numcodecs:abc/codec interface from a WebAssembly component instance.

The ctx must refer to the same store in which the instance was instantiated.

Warning

The WasmCodecComponent does not own the provided instance and never calls [Instance::drop]. It is the responsibility of the code creating the WasmCodecComponent to destroy the instance after the component, and all codecs created from it, have been destroyed.

Errors

Errors if the instance does not export the numcodecs:abc/codec interface or if interacting with the component fails.

impl WasmCodecComponent

Methods for implementing the DynCodecType trait

pub fn codec_id(&self) -> &str

Codec identifier.

pub fn codec_config_schema(&self) -> &Schema

JSON schema for the codec’s configuration.

pub fn codec_from_config<'de, D: Deserializer<'de>>( &self, ctx: impl AsContextMut, config: D, ) -> Result<WasmCodec, D::Error>

Instantiate a codec of this type from a serialized configuration.

The config must not contain an id field. If the config may contain one, use the codec_from_config_with_id helper function.

The config must be compatible with JSON encoding.

Errors

Errors if constructing the codec or interacting with the component fails.